automatik push

+++ Cpp/H/IO.h
@@ -2,9 +2,9 @@
 #include "typedefs.h"
 
 #define PIC1_COMMAND 0x20
-#define PIC1_DATA 0x21
+#define PIC1_DATA (uint16_t)0x21
 #define PIC2_COMMAND 0xA0
-#define PIC2_DATA 0xA1
+#define PIC2_DATA (uint16_t)0xA1
 
 #define ICW1_INIT 0x10
 #define ICW1_ICW4 0x01
